+++ lib/constants/preference_keys.dart
@@ -1,0 +1,4 @@
+class PreferenceKeys{
+  // define shared preference keys here
+  static String isFirstRun = 'isFirstRun';
+}
