+++ lib/screens/onboarding/onboarding_screen.dart
@@ -1,7 +1,7 @@
 import 'package:carousel_slider/carousel_slider.dart';
 import 'package:flutter/material.dart';
-import 'package:partypal/configs/asset_paths.dart';
-import 'package:partypal/configs/route_paths.dart';
+import 'package:partypal/constants/asset_paths.dart';
+import 'package:partypal/constants/route_paths.dart';
 import 'package:partypal/configs/router_config.dart';
 import 'package:partypal/widgets/onboarding_card.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
